implemented browser-router in the App component. Moved app component code to separate home page.
