+++ package.json
@@ -4,7 +4,8 @@
     "firebase": "^10.8.0",
     "firebase-tools": "^13.2.1",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.22.0"
   },
   "devDependencies": {
     "@vitejs/plugin-react": "^4.2.1",
