+++ src/main.jsx
@@ -4,7 +4,5 @@ import App from "./App.jsx";
 import "./global.css";
 
 ReactDOM.createRoot(document.getElementById("root")).render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>
+  <App />
 );
