+++ src/pages/Payments.jsx
@@ -1,0 +1,8 @@
+import React from 'react'
+
+export default function Payments() {
+
+    return (
+        <h1>Payment page</h1>
+    )
+}
